Add files via upload
    1.Enhanced individual privacy verifiable on the blockchain
    2.Flexible identity solutions for pseudonymity and self asserted identities
    3.Build on a decentralized web of trust

Privacy:

Defensive security is a multi-faceted concern to individuals, requiring solutions that can address multiple challenges. BlockchainMe is a tool to promote the encryption layer of cryptocurrency, to help prevent skimming, catfishing, counterfeiting and other fraudulent activity are critical.

Verifiable:

BlockchainMe utilizes proof-of-existence, your information is never sent or stored centrally on any database. You own your information. Your Id is only verifiable by the people you chose to give it to. This is a voluntary identification service that adapts to the increasingly diversified needs for contextual identity.

Identity:

BlockchainMe offers counterfeit-resistant and tamper-proof identification credentials that hold the only the information you choose. Use your name or p